+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update tests to run against python 3.8 and 3.9
 - add collection_of
 - add users.can
+- add uuid to browse table
 
 ### Added
 ## [6.14.0] - 2020-12-21